Add authentication and Exception handlers

+++ src/main/java/com/userdatabase/userdemo/service/UserService.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 
 import com.userdatabase.userdemo.database.UserRepository;
 import com.userdatabase.userdemo.entity.User;
+import com.userdatabase.userdemo.exception.UserException;
 
 @Service
 public class UserService {
@@ -14,22 +15,33 @@ public class UserService {
 	@Autowired
 	UserRepository repository;
 
-	public Optional<User> findUser(String userId) {
-		Optional<User> user = repository.findById(userId);
-		return user;
-	}
-
 	public void saveUser(User user) {
 		repository.save(user);
 	}
 
-	public void deleteUser(String userId) {
+	public User findUser(String userId) throws UserException {
+
+		// Exception for if there is no user with given User Id
+		Optional<User> user = repository.findById(userId);
+		if (user == null) {
+			throw new UserException("No such record found for User id :" + userId);
+		}
+
+		return user.get();
+	}
+
+	public void deleteUser(String userId) throws UserException {
+		User user = findUser(userId);
+
+		// Exception for if the person trying to delete information is not valid
+		if (!(isUserValid(user, user.getName(), user.getPassword()))) {
+			throw new UserException("You are not valid!");
+		}
 		repository.deleteById(userId);
 	}
 
-	// TODO Make password variable, change code
-	public boolean validateUser(String user, String password) {
-		if (user.equals("Ira") && password.equals("mypassword")) {
+	public boolean isUserValid(User user, String userName, String userPassword) {
+		if (user.getName().equals(userName) && user.getPassword().equals(userPassword)) {
 			return true;
 		}
 		return false;
